updates to DELETE method to make sure we send back a happy response if it exists
updates to DELETE method. break when we get a succuessful result and end the request outside of the for loop. otherwise we send a response back on the first iteration. also, use 204 since we're not sending back any data

all the whitespace and tabs

updates to DELETE method. break when we get a succuessful result and end the request outside of the for loop. otherwise we send a response back on the first iteration. also, use 204 since we're not sending back any data

iterate through all the things and send a happy resposne if it exists. otherwise sad response